// src/app/models/event.ts
export interface EventResponse {
    perfilImage: {
        createdAt?: string,
        key?: string,
        name?: string,
        size?: number,
        url: string
        __v?: string,
        _id?: string
    }
    _id: string;
    crp: string;
    start: string;
    end: string;
    title: string;
    time: string;
    cpfCliente: string;
    mes: string;
    ano: string;
    dia: string;
    realizado: boolean;
    pago: boolean;
    valorConsulta: string;
    data: string;
    meet: string;
    color: string;
    __v: number;
}